Add scripts to launch simulations with the new geometry in Comsol
This Gap simulation is defined with just 2 ring in the field cage. The
Anode and Gate mesh are defined with hexagons and wire, respetively,
only in the central region. The rest is set as a solid plate.
We have readjust the cathode voltate to mantain the 5kV drift in the
original simulation, and so the eletric field.